requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
beautifulsoup4==4.9.3
certifi==2020.6.20
chardet==3.0.4
discord.py==1.5.1
idna==2.10
lxml==4.6.0
multidict==4.7.6
pika==1.1.0
python-dotenv==0.15.0
pyzmq==20.0.0
selenium==3.141.0
soupsieve==2.0.1
urllib3==1.25.10
yarl==1.5.1
